+++ lib/telnet.c
@@ -1512,7 +1512,8 @@ static CURLcode telnet_do(struct Curl_easy *data, bool *done)
           keepon = FALSE;
           /* FIXME: in test 1452, macOS sees a ECONNRESET sometimes?
            * Is this the telnet test server not shutting down the socket
-           * in a clean way? */
+           * in a clean way? Seems to be timing related, happens more
+           * on slow debug build */
           if(data->state.os_errno == ECONNRESET) {
             DEBUGF(infof(data, "telnet_do(handle=%p), unexpected ECONNRESET"
                          " on recv", data));
